chore: move code in individual files
